Moved inference error handling into errors.py.
