--- src/sagemaker_inference/error_handling.py
@@ -1,44 +1,0 @@
-# Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the 'License'). You
-# may not use this file except in compliance with the License. A copy of
-# the License is located at
-#
-#     http://aws.amazon.com/apache2.0/
-#
-# or in the 'license' file accompanying this file. This file is
-# distributed on an 'AS IS' BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
-# ANY KIND, either express or implied. See the License for the specific
-# language governing permissions and limitations under the License.
-
-
-class BaseInferenceToolkitError(Exception):
-    """Exception used to indicate a problem that occurred during inference.
-
-    This is meant to be extended from so that customers may handle errors within inference servers.
-
-    :param status_code: HTTP Error Status Code to send to client
-    :param message: Response message to send to client
-    :param phrase: Response body to send to client
-    """
-    def __init__(self, status_code, message, phrase):
-        self.status_code = status_code
-        self.message = message
-        self.phrase = phrase
-
-
-class GenericInferenceToolkitError(BaseInferenceToolkitError):
-    """Exception used to indicate a problem that occurred during inference.
-
-    This is meant to be a generic implementation of the BaseInferenceToolkitError for re-raising unexpected
-    exceptions in a way that can be sent back to the client.
-
-    :param status_code: HTTP Error Status Code to send to client
-    :param message: Response message to send to client
-    :param phrase: Response body to send to client
-    """
-
-    def __init__(self, status_code, message=None, phrase=None):
-        self.status_code = status_code
-        self.message = message if message else "Invalid Request"
-        self.phrase = phrase if phrase else message
